+++ ptrlv/src/main/java/com/cc/listview/ptrlv/TXPtrListView.java
@@ -158,7 +158,7 @@ public class TXPtrListView<T> extends TXPTRAndLMBase<T> {
     }
 
     @Override
-    public void loadError(int errorCode, String message) {
+    public void loadError(long errorCode, String message) {
         mAdapter.loadError(errorCode, message);
     }
 
@@ -238,7 +238,7 @@ public class TXPtrListView<T> extends TXPTRAndLMBase<T> {
         }
 
         @Override
-        public View getErrorView(ViewGroup parent, int errorCode, String message) {
+        public View getErrorView(ViewGroup parent, long errorCode, String message) {
             View view = LayoutInflater.from(parent.getContext()).inflate(listView.getErrorLayoutId(), null);
 
             TextView tv = (TextView) view.findViewById(R.id.tx_ids_list_error_msg);
@@ -265,7 +265,7 @@ public class TXPtrListView<T> extends TXPTRAndLMBase<T> {
         }
 
         @Override
-        public View getLoadMoreErrorView(ViewGroup parent, int errorCode, String message) {
+        public View getLoadMoreErrorView(ViewGroup parent, long errorCode, String message) {
             View view = LayoutInflater.from(parent.getContext()).inflate(listView.getLoadMoreErrorLayoutId(), parent, false);
 
             view.findViewById(R.id.tx_ids_list_load_more_error).setOnClickListener(new OnClickListener() {
